+++ vcd/resource_vcd_catalog_media_test.go
@@ -4,11 +4,11 @@ package vcd
 
 import (
 	"fmt"
+	"regexp"
 	"testing"
 
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
-	"github.com/vmware/go-vcloud-director/v2/govcd"
 )
 
 var TestAccVcdCatalogMedia = "TestAccVcdCatalogMediaBasic"
@@ -16,10 +16,9 @@ var TestAccVcdCatalogMediaDescription = "TestAccVcdCatalogMediaBasicDescription"
 
 func TestAccVcdCatalogMediaBasic(t *testing.T) {
 
-	var catalogItem govcd.CatalogItem
 	var params = StringMap{
 		"Org":              testConfig.VCD.Org,
-		"Catalog":          testSuiteCatalogName,
+		"Catalog":          testConfig.VCD.Catalog.Name,
 		"CatalogMediaName": TestAccVcdCatalogMedia,
 		"Description":      TestAccVcdCatalogMediaDescription,
 		"MediaPath":        testConfig.Media.MediaPath,
@@ -45,7 +44,7 @@ func TestAccVcdCatalogMediaBasic(t *testing.T) {
 			resource.TestStep{
 				Config: configText,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckVcdCatalogMediaExists("vcd_catalog_media."+TestAccVcdCatalogMedia, &catalogItem),
+					testAccCheckVcdCatalogMediaExists("vcd_catalog_media."+TestAccVcdCatalogMedia),
 					resource.TestCheckResourceAttr(
 						"vcd_catalog_media."+TestAccVcdCatalogMedia, "name", TestAccVcdCatalogMedia),
 					resource.TestCheckResourceAttr(
@@ -54,12 +53,17 @@ func TestAccVcdCatalogMediaBasic(t *testing.T) {
 						"vcd_catalog_media."+TestAccVcdCatalogMedia, "metadata.mediaItem_metadata", "mediaItem Metadata"),
 					resource.TestCheckResourceAttr(
 						"vcd_catalog_media."+TestAccVcdCatalogMedia, "metadata.mediaItem_metadata2", "mediaItem Metadata2"),
+					resource.TestMatchOutput("owner_name", regexp.MustCompile(`^\S+`)),
+					resource.TestMatchOutput("creation_date", regexp.MustCompile(`^^\d{4}-\d{2}-\d{2}.*`)),
+					resource.TestCheckOutput("status", "RESOLVED"),
+					resource.TestMatchOutput("storage_profile_name", regexp.MustCompile(`^\S+`)),
+					testCheckMediaNonStringOutputs(),
 				),
 			},
 			resource.TestStep{
 				Config: updateConfigText,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckVcdCatalogMediaExists("vcd_catalog_media."+TestAccVcdCatalogMedia, &catalogItem),
+					testAccCheckVcdCatalogMediaExists("vcd_catalog_media."+TestAccVcdCatalogMedia),
 					resource.TestCheckResourceAttr(
 						"vcd_catalog_media."+TestAccVcdCatalogMedia, "name", TestAccVcdCatalogMedia),
 					resource.TestCheckResourceAttr(
@@ -72,11 +76,39 @@ func TestAccVcdCatalogMediaBasic(t *testing.T) {
 						"vcd_catalog_media."+TestAccVcdCatalogMedia, "metadata.mediaItem_metadata3", "mediaItem Metadata3"),
 				),
 			},
+			resource.TestStep{
+				ResourceName:      "vcd_catalog_media." + TestAccVcdCatalogMedia + "-import",
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateIdFunc: importStateIdOrgCatalogObject(testConfig, TestAccVcdCatalogMedia),
+				// These fields can't be retrieved from catalog media data
+				ImportStateVerifyIgnore: []string{"media_path", "upload_piece_size", "show_upload_progress"},
+			},
 		},
 	})
 }
 
-func testAccCheckVcdCatalogMediaExists(mediaName string, catalogItem *govcd.CatalogItem) resource.TestCheckFunc {
+func testCheckMediaNonStringOutputs() resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		outputs := s.RootModule().Outputs
+
+		if outputs["is_iso"].Value != true {
+			return fmt.Errorf("is_iso value didn't match")
+		}
+
+		if outputs["is_published"].Value != false {
+			return fmt.Errorf("is_published value didn't match")
+		}
+
+		if regexp.MustCompile(`^\d+$`).MatchString(fmt.Sprintf("%s", outputs["size"].Value)) {
+			return fmt.Errorf("size value isn't int")
+		}
+
+		return nil
+	}
+}
+
+func testAccCheckVcdCatalogMediaExists(mediaName string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		catalogMediaRs, ok := s.RootModule().Resources[mediaName]
 		if !ok {
@@ -94,17 +126,16 @@ func testAccCheckVcdCatalogMediaExists(mediaName string, catalogItem *govcd.Cata
 			return fmt.Errorf(errorRetrievingOrg, testConfig.VCD.Org+" and error: "+err.Error())
 		}
 
-		catalog, err := adminOrg.GetCatalogByName(testSuiteCatalogName, false)
+		catalog, err := adminOrg.GetCatalogByName(testConfig.VCD.Catalog.Name, false)
 		if err != nil {
-			return fmt.Errorf("catalog %s does not exist (%s)", testSuiteCatalogName, err)
+			return fmt.Errorf("catalog %s does not exist (%s)", testConfig.VCD.Catalog.Name, err)
 		}
 
-		newCatalogItem, err := catalog.GetCatalogItemByName(catalogMediaRs.Primary.Attributes["name"], false)
+		foundMedia, err := catalog.GetMediaByName(catalogMediaRs.Primary.Attributes["name"], false)
 		if err != nil {
-			return fmt.Errorf("catalog media %s does not exist (%#v)", catalogMediaRs.Primary.ID, catalogItem.CatalogItem)
+			return fmt.Errorf("catalog media %s does not exist (%#v)", catalogMediaRs.Primary.ID, foundMedia.Media)
 		}
 
-		catalogItem = newCatalogItem
 		return nil
 	}
 }
@@ -121,13 +152,13 @@ func testAccCheckCatalogMediaDestroy(s *terraform.State) error {
 			return fmt.Errorf(errorRetrievingOrg, testConfig.VCD.Org+" and error: "+err.Error())
 		}
 
-		catalog, err := adminOrg.GetCatalogByName(testSuiteCatalogName, false)
+		catalog, err := adminOrg.GetCatalogByName(testConfig.VCD.Catalog.Name, false)
 		if err != nil {
 			return fmt.Errorf("catalog query %s ended with error: %#v", rs.Primary.ID, err)
 		}
 
 		mediaName := rs.Primary.Attributes["name"]
-		_, err = catalog.GetCatalogItemByName(mediaName, false)
+		_, err = catalog.GetMediaByName(mediaName, false)
 
 		if err == nil {
 			return fmt.Errorf("catalog media %s still exists", mediaName)
@@ -153,7 +184,35 @@ const testAccCheckVcdCatalogMediaBasic = `
     mediaItem_metadata2 = "mediaItem Metadata2"
   }
 }
-`
+
+output "creation_date" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.creation_date
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "is_iso" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.is_iso
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "owner_name" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.owner_name
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "is_published" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.is_published
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "size" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.size
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "status" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.status
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}
+output "storage_profile_name" {
+  value = vcd_catalog_media.{{.CatalogMediaName}}.storage_profile_name
+  depends_on = [vcd_catalog_media.{{.CatalogMediaName}}]
+}`
 
 const testAccCheckVcdCatalogMediaUpdate = `
   resource "vcd_catalog_media"  "{{.CatalogMediaName}}" {